tests: extract random stress tests into dedicated file
