+++ tests/test_harness.h
@@ -17,6 +17,8 @@ void fini_test(void);
 void lock_snapshot_take(lock_snapshot_t *snap);
 void lock_snapshot_assert_delta_one(const lock_snapshot_t *before);
 void test_lock_assert_unheld(void);
+void check_stats(int verbose);
+void print_final_stats(void);
 
 tealet_t *tealet_new_native_call(tealet_t *m, tealet_run_t run, void **parg, void *stack_far);
 int tealet_test_new_dispatch(tealet_t *t, tealet_t **out, tealet_run_t run, void **parg, void *stack_far);
