+++ tests/test_stress.h
@@ -1,0 +1,7 @@
+#ifndef TEST_STRESS_H
+#define TEST_STRESS_H
+
+void test_random(void);
+void test_random2(void);
+
+#endif
